Extract common SEV code
Also creates common module for other code shared between
KBCs.
