+++ src/common/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) 2022 IBM
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+pub mod sev;
